+++ .eslintrc.js
@@ -4,7 +4,8 @@ module.exports = {
     es2020: true
   },
   extends: [
-    'standard'
+    'standard',
+    'plugin:@typescript-eslint/recommended'
   ],
   parser: '@typescript-eslint/parser',
   parserOptions: {